Adds integration tests, updates Db settings (#7, #14)
