+++ src/test/java/com/naeemark/fbs/services/AccountServiceTest.java
@@ -8,13 +8,11 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.MockitoAnnotations;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import static com.naeemark.fbs.utils.Constants.ERROR_ACCOUNT_NOT_FOUND;
@@ -28,22 +26,20 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest
-@Transactional
-@ActiveProfiles("test")
+@ExtendWith(MockitoExtension.class)
 class AccountServiceTest {
 
-    @Autowired
+    @InjectMocks
     private AccountServiceImpl service;
 
-    @MockBean
+    @Mock
     private AccountRepository repository;
 
-    final Account testAccount = new Account(1, 1);
+    private Account testAccount ;
 
     @BeforeEach
-    void mockCache() {
-        MockitoAnnotations.initMocks(this);
+    public void setup(){
+        testAccount = new Account(1, 1);
     }
 
     @Test
